backend/requirements.txt
Django>=5.1.4,<5.3
djangorestframework>=3.15.2,<3.16
drf-spectacular>=0.27,<0.28
Markdown>=3.7,<4
gunicorn>=23.0.0,<24.0.0
django-cors-headers>=4.6.0,<4.7
djangorestframework-simplejwt>=5.3.1,<5.4
django-filter>=24.3,<25
django-allauth>=64.2,<65
dj-rest-auth>=6.0.0,<6.1
PyJWT>=2.9.0,<3
cryptography>=42.0.0,<43
requests>=2.32.3,<2.33
django-import-export>=4.3,<4.4
python-dotenv>=1.0.1,<1.1
psycopg2-binary>=2.9,<2.10
django-storages==1.14.6
boto3>=1.35,<1.36
pillow==11.3.0
stripe>=11.1.0,<12
